day 2 zig: refactor to make test possible.
was stuck for the longest time on a shitty compiler error:

    error: expected type 'type', found 'fn(comptime type) type'

raised by the following line:

    const lines: std.mem.SplitIterator = std.mem.split(u8, in, n);

the problem here is that SplitIterator is a generic function
which comptime builds a type.

The shitty thing about it: it was shown as the error on this line:

    pub fn day02(in: []const u8) !common.Answer {

which preceded the real error line.